🧰: move `text-editor.js` into `text` directory
